crash fix
changed logical lines into numerical lines

+++ magia_medfilter_slice.m
@@ -4,7 +4,7 @@ old_slice = squeeze(old_slice);
 new_slice = old_slice;
 
 for x = 1:size(new_slice,1)
-    old_line = old_slice(x,:);
+    old_line = double(old_slice(x,:));
     new_line = medfilt1(old_line,3);
     new_slice(x,:) = new_line;
 end
